+++ src/components/LogoAndNotification.tsx
@@ -1,20 +1,15 @@
 import Logo from '@assets/images/roomit_logo.png';
-import { BASE_URL } from '@constants/constants';
-import useAuthStore from '@store/authStore';
-import { getAuthToken, getRole } from '@utils/auth';
-import { NativeEventSource, EventSourcePolyfill } from 'event-source-polyfill';
-import { useEffect, useState } from 'react';
+import { getRole } from '@utils/auth';
 import { RiNotification3Line } from 'react-icons/ri';
 import { Link, useNavigate } from 'react-router-dom';
-import { SseAlarm } from '@typings/types';
-import NotiContainer from './notiContainer';
 
-const LogoAndNotification = () => {
+interface HeaderProps {
+  isLogin: boolean;
+}
+
+const LogoAndNotification = ({ isLogin }: HeaderProps) => {
   const navigate = useNavigate();
-  const { isLogin } = useAuthStore();
   const role = getRole();
-
-  const [message, setMessage] = useState<SseAlarm>();
 
   const handleMoveToNotiPageClick = () => {
     if (role === 'ROLE_USER') {
@@ -23,34 +18,6 @@ const LogoAndNotification = () => {
       navigate('/host-noti');
     }
   };
-
-  useEffect(() => {
-    console.log('useEffect');
-    const EventSource = EventSourcePolyfill || NativeEventSource;
-    const token = getAuthToken() || '';
-    console.log(token);
-    const eventSource = new EventSource(`${BASE_URL}/api/subscribe`, {
-      headers: {
-        Authorization: token,
-      },
-      withCredentials: true,
-    });
-
-    console.log('EventSource initialized:', eventSource);
-
-    eventSource.onmessage = (event) => {
-      const newMessage = event.data;
-      setMessage(newMessage);
-    };
-
-    eventSource.onerror = (error) => {
-      console.log(error);
-    };
-
-    return () => {
-      eventSource.close();
-    };
-  }, []);
 
   return (
     <>
@@ -67,7 +34,6 @@ const LogoAndNotification = () => {
           onClick={handleMoveToNotiPageClick}
         />
       )}
-      {isLogin && message && <NotiContainer message={message} />}
     </>
   );
 };
