+++ src/components/notiContainer.tsx
@@ -1,7 +1,13 @@
 import { SseAlarm } from '@typings/types';
+import { AiFillNotification } from 'react-icons/ai';
 
 const NotiContainer = ({ message }: { message: SseAlarm }) => {
-  return <div className='h-[50px] w-custom bg-primary'>{message.content}</div>;
+  return (
+    <div className='fixed top-8 z-[2000] flex h-[70px] w-custom items-center justify-start gap-3 rounded-lg bg-white px-4 py-2 text-base text-black shadow-custom'>
+      <AiFillNotification className='size-5 text-primary' />
+      {message.content}
+    </div>
+  );
 };
 
 export default NotiContainer;
